+++ scripts/render_help.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+# Render the help to an SVG; render that SVG to a .png image
+
+RENDERED_FILE_BASE=yaralyzer_help
+RENDERED_SVG="$RENDERED_FILE_BASE.svg"
+RENDERED_PNG="doc/rendered_images/$RENDERED_FILE_BASE.png"
+
+scripts/render_help.py
+inkscape --export-filename="$RENDERED_PNG" "$RENDERED_SVG"
+rm $RENDERED_SVG