+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next Release
+# 0.3.0
 * Add `--rules-dir` option for loading all `.yara` files in directories
 * Add `--regex_modifier` option